+++ core/src/rest/stock/intraday/mod.rs
@@ -2,6 +2,7 @@
 
 mod candles;
 mod quote;
+mod quotes;
 mod ticker;
 mod tickers;
 mod trades;
@@ -9,6 +10,7 @@ mod volumes;
 
 pub use candles::CandlesRequestBuilder;
 pub use quote::QuoteRequestBuilder;
+pub use quotes::QuotesRequestBuilder;
 pub use ticker::TickerRequestBuilder;
 pub use tickers::TickersRequestBuilder;
 pub use trades::TradesRequestBuilder;